# Faster map extract benchmark