--- a/.tool-versions
+++ b/.tool-versions
@@ -1,1 +1,1 @@
-elixir 1.5.0-rc.0
+elixir 1.11.2