--- a/mix.exs
+++ b/mix.exs
@@ -35,7 +35,8 @@ defmodule Extractable.Mixfile do
   # Type "mix help deps" for more examples and options
   defp deps do
     [
-      {:ex_doc, "~> 0.14", only: :dev}
+      {:ex_doc, "~> 0.14", only: :dev},
+      {:benchee, "~> 1.0", only: :dev}
     ]
   end