Related to #13, this is a benchmark to check the performance of different implementations of `extract` for maps (which might also be applicable for MapSets).

Running this benchmark on my machine yields the following:

```
$ mix run benchmarks/map_extract.exs 
Operating System: Linux
CPU Information: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
Number of Available Cores: 8
Available memory: 7.61 GB
Elixir 1.11.2
Erlang 23.0

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 1 s
memory time: 1 s
parallel: 1
inputs: 10, 100, empty
Estimated total run time: 48 s

Benchmarking MapIterator with input 10...
Benchmarking MapIterator with input 100...
Benchmarking MapIterator with input empty...
Benchmarking MapKeys with input 10...
Benchmarking MapKeys with input 100...
Benchmarking MapKeys with input empty...
Benchmarking MapKeysSize with input 10...
Benchmarking MapKeysSize with input 100...
Benchmarking MapKeysSize with input empty...
Benchmarking MapList with input 10...
Benchmarking MapList with input 100...
Benchmarking MapList with input empty...

##### With input 10 #####
Name                  ips        average  deviation         median         99th %
MapKeysSize       12.34 M       81.01 ns ±10835.42%           0 ns         296 ns
MapList           10.64 M       94.00 ns ±13658.06%           0 ns         368 ns
MapKeys           10.44 M       95.83 ns ±19165.13%           0 ns         268 ns
MapIterator        1.77 M      564.74 ns  ±2266.35%         471 ns        1001 ns

Comparison: 
MapKeysSize       12.34 M
MapList           10.64 M - 1.16x slower +12.99 ns
MapKeys           10.44 M - 1.18x slower +14.81 ns
MapIterator        1.77 M - 6.97x slower +483.72 ns

Memory usage statistics:

Name           Memory usage
MapKeysSize           464 B
MapList               680 B - 1.47x memory usage +216 B
MapKeys               464 B - 1.00x memory usage +0 B
MapIterator          1032 B - 2.22x memory usage +568 B

**All measurements for memory usage were the same**

##### With input 100 #####
Name                  ips        average  deviation         median         99th %
MapKeysSize      853.10 K        1.17 μs   ±484.54%        1.08 μs        2.68 μs
MapKeys          844.04 K        1.18 μs   ±518.66%        1.08 μs        2.84 μs
MapList          119.24 K        8.39 μs    ±39.82%        8.08 μs       12.94 μs
MapIterator       72.67 K       13.76 μs    ±22.55%       13.18 μs       19.22 μs

Comparison: 
MapKeysSize      853.10 K
MapKeys          844.04 K - 1.01x slower +0.0126 μs
MapList          119.24 K - 7.15x slower +7.21 μs
MapIterator       72.67 K - 11.74x slower +12.59 μs

Memory usage statistics:

Name           Memory usage
MapKeysSize         1.67 KB
MapKeys             1.67 KB - 1.00x memory usage +0 KB
MapList             4.13 KB - 2.47x memory usage +2.45 KB
MapIterator         7.85 KB - 4.70x memory usage +6.18 KB

**All measurements for memory usage were the same**

##### With input empty #####
Name                  ips        average  deviation         median         99th %
MapList            1.93 B        0.52 ns  ±6854.19%           0 ns           0 ns
MapKeysSize        1.01 B        0.99 ns ±29150.80%           0 ns           0 ns
MapKeys            0.40 B        2.50 ns ±51753.43%           0 ns           0 ns
MapIterator      0.0483 B       20.71 ns ±55288.94%           0 ns           0 ns

Comparison: 
MapList            1.93 B
MapKeysSize        1.01 B - 1.90x slower +0.47 ns
MapKeys            0.40 B - 4.81x slower +1.98 ns
MapIterator      0.0483 B - 39.94x slower +20.19 ns

Memory usage statistics:

Name           Memory usage
MapList                 0 B
MapKeysSize             0 B - 1.00x memory usage +0 B
MapKeys                 0 B - 1.00x memory usage +0 B
MapIterator            16 B - ∞ x memory usage +16 B

**All measurements for memory usage were the same**

```

This clearly seems to indicate that a `Map.keys` with extra `map_size` check is the preferred one based on time and memory efficiency.